Move OrderUpdater to Spree::Order::Updater
